requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
pytest-django==4.5.2
pytest-testdox==3.0.1
pytz==2023.3
sqlparse==0.4.3
